+++ CEphProject/src/containers/map.h
@@ -15,15 +15,20 @@
 
 
 
+#include <stddef.h>
+
+
+
 typedef struct map map;
+
 
 
 map* map_create(void);
 void map_destroy(map* m);
-void* map_search(map* m, int key);
-void map_insert(map* m, int key, void* value);
-void map_erase(map* m, int key);
-void map_for_each_item(map* m, void(*func)(int, void*));
+void* map_search(map* m, size_t key);
+void map_insert(map* m, size_t key, void* value);
+void map_erase(map* m, size_t key);
+void map_for_each_item(map* m, void(*func)(size_t, void*));
 int map_get_size(map* m);
 
 #endif /* !MAP_H */
